Move devices back into backstore module
