+++ src/engine/strat_engine/backstore/blockdevmgr.rs
@@ -17,11 +17,13 @@ use devicemapper::{Bytes, Device, LinearDevTargetParams, LinearTargetParams, Sec
 use crate::{
     engine::{
         strat_engine::{
-            backstore::{blockdev::StratBlockDev, initialize::initialize_devices},
+            backstore::{
+                blockdev::StratBlockDev, devices::process_and_verify_devices,
+                initialize::initialize_devices,
+            },
             crypt::CryptHandle,
             metadata::MDADataSize,
             names::KeyDescription,
-            raw_devices::process_and_verify_devices,
             serde_structs::{BaseBlockDevSave, BaseDevSave, Recordable},
         },
         types::{DevUuid, PoolUuid},
