+++ src/engine/strat_engine/backstore/mod.rs
@@ -8,6 +8,7 @@ mod blockdev;
 mod blockdevmgr;
 mod cache_tier;
 mod data_tier;
+mod devices;
 mod initialize;
 mod range_alloc;
 mod shared;
@@ -15,4 +16,4 @@ mod shared;
 pub use self::{backstore::Backstore, blockdev::StratBlockDev};
 
 #[cfg(test)]
-pub use self::initialize::initialize_devices;
+pub use self::{devices::process_and_verify_devices, initialize::initialize_devices};
