+++ src/engine/strat_engine/raw_devices/mod.rs
@@ -2,10 +2,6 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-mod devices;
 mod identify;
 
-pub use self::{
-    devices::{process_and_verify_devices, InitDeviceInfo},
-    identify::{find_all, identify_block_device, DeviceInfo, LuksInfo, StratisInfo},
-};
+pub use self::identify::{find_all, identify_block_device, DeviceInfo, LuksInfo, StratisInfo};
